fixing docs failure
